+++ R/data.R
@@ -1,33 +1,3 @@
-#' COG 2023
-#'
-#' A dataset containing insee code and wording of commune as of January 1, 2023. COG mean
-#' Code Officiel Géographique
-#'
-
-#' @format ## `cog_2023`
-#' A data frame with 34990 rows and 2 columns:
-#' \describe{
-#'   \item{COM}{insee code}
-#'   \item{LIBELLE}{Name of commune}
-#' }
-#' @source \url{https://www.insee.fr/fr/information/6800675}
-"cog_2023"
-
-#' COG 2024
-#'
-#' A dataset containing insee code and wording of commune as of January 1, 2024. COG mean
-#' Code Officiel Géographique
-#'
-
-#' @format ## `com_2024`
-#' A data frame with 34980 rows and 2 columns:
-#' \describe{
-#'   \item{COM}{insee code}
-#'   \item{LIBELLE}{Name of commune}
-#' }
-#' @source \url{https://www.insee.fr/fr/information/7766585}
-"com_2024"
-
 #' French Communes Table (2025)
 #'
 #' Data for French communes from the INSEE file "v_commune_2025.csv".